% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_fct_clean_tidy_data.R
\name{clean_tidy_data}
\alias{clean_tidy_data}
\title{Get market data which has gone through assertR cleaning}
\usage{
clean_tidy_data()
}
\value{
Return a dataframe which has gone through assertr cleaning
}
\description{
Get market data which has gone through assertR cleaning
}
\examples{
\dontrun{
clean_tidy_data()
}
}
