% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_fct_get_data.R
\name{get_data}
\alias{get_data}
\title{Read data from online only if internet available or if data is stale}
\usage{
get_data()
}
\value{
a dataframe of agricultural commodity prices in raw form
}
\description{
Read data from online only if internet available or if data is stale
}
\examples{
\dontrun{
get_data()
}
}
