% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_utils_db_data.R
\name{get_districts}
\alias{get_districts}
\title{Get district details for each participating district}
\usage{
get_districts()
}
\value{
Returns a dataframe of participating districts
}
\description{
Get district details for each participating district
}
\examples{
\dontrun{
get_districts
}
}
