% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_utils_db_data.R
\name{get_extension_officers}
\alias{get_extension_officers}
\title{Get extension officer details for each participating ward}
\usage{
get_extension_officers()
}
\value{
Returns a dataframe of ward extension officers
}
\description{
Get extension officer details for each participating ward
}
\examples{
\dontrun{
get_extension_officers()
}
}
