% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_fct_get_raw_data.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Gets raw data from data collected on a form hosted on Kobo server}
\usage{
get_raw_data()
}
\value{
returns a dataframe of the survey data and sets it as a global variable
named `survey`
}
\description{
Gets raw data from data collected on a form hosted on a Kobo server.
 Requires that your local environment variables for `sakirp_user` 
 (a Kobo user account) and `sakirp_pw` (the corresponding password) be set using
 `Sys.setenv(sakirp_user = "kobo user name,sakirp_pw = "kobo password`
}
\examples{
\dontrun{
get_raw_data()
}
}
