% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_utils_db_data.R
\name{get_wards}
\alias{get_wards}
\title{Get ward details for each participating ward}
\usage{
get_wards()
}
\value{
Returns a dataframe of participating wards
}
\description{
Get ward details for each participating ward
}
\examples{
\dontrun{
get_wards()
}
}
