% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_utils_db_data.R
\name{postgres_connection}
\alias{postgres_connection}
\title{Get a postrgres connection object}
\usage{
postgres_connection()
}
\value{
A postgres connection object
}
\description{
Get a postrgres connection object. An ODBC connection has to have been 
created on the local machine running the application
}
\examples{
\dontrun{
postgres_connection()
}
}
