% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_fct_price_dygraph.R
\name{price_dygraph}
\alias{price_dygraph}
\title{Create a timeseries graph of commodity prices}
\usage{
price_dygraph(data, product, dstrct, wrd)
}
\arguments{
\item{data}{Price data}

\item{product}{Product selected by user}

\item{dstrct}{District of interest}

\item{wrd}{Ward of interest}
}
\value{
a dypgraph object
}
\description{
Create a timeseries graph of commodity prices
}
\examples{
\dontrun{
 price_dygraph(price_long(),product = "Kigoma Njano",
 dstrct = "Kigoma",wrd = "Kalinzi")
}
}
