% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_fct_price_long.R
\name{price_long}
\alias{price_long}
\title{Agricultural commodity price data in long format}
\usage{
price_long()
}
\value{
Returns a long dataframe of agricultural commodities and their prices
}
\description{
Agricultural commodity price data in long format
}
\examples{
price_long
}
