% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_fct_price_wide.R
\name{price_wide}
\alias{price_wide}
\title{Agricultural commodity price data in wide format}
\usage{
price_wide()
}
\value{
Returns a wide dataframe of agricultural commodities and their prices
}
\description{
Agricultural commodity price data in wide format
}
\examples{
price_wide()
}
