% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_market_fct_tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\title{Get selected columns of market data}
\usage{
tidy_data()
}
\value{
Returns a dataframe of select columns
}
\description{
Get selected columns of market data
}
\examples{
\dontrun{
tidy_data()
}
}
